+++ lib/segment/src/index/field_index/field_index_base/field_index.rs
@@ -10,8 +10,8 @@ use super::value_indexer::ValueIndexer;
 use crate::common::Flusher;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::bool_index::{BoolIndex, BoolIndexRead};
-use crate::index::field_index::full_text_index::read_ops::FullTextIndexRead;
-use crate::index::field_index::full_text_index::text_index::FullTextIndex;
+use crate::index::field_index::full_text_index::FullTextIndex;
+use crate::index::field_index::full_text_index::full_text_index_read::FullTextIndexRead;
 use crate::index::field_index::geo_index::{GeoMapIndex, GeoMapIndexRead};
 use crate::index::field_index::map_index::MapIndex;
 use crate::index::field_index::null_index::{NullIndex, NullIndexRead};
