+++ lib/segment/src/index/field_index/full_text_index/read_ops.rs
@@ -1,235 +1,215 @@
-use std::borrow::Cow;
-
-use ahash::AHashMap;
 use common::counter::hardware_accumulator::HwMeasurementAcc;
 use common::counter::hardware_counter::HardwareCounterCell;
-use common::iterator_ext::IteratorExt;
 use common::types::PointOffsetType;
 use common::universal_io::UserData;
+use serde_json::Value;
 
-use super::inverted_index::{Document, ParsedQuery, TokenId, TokenSet};
-use super::text_index::PayloadMatchQueryType;
-use super::tokenizers::{Tokenizer, TokenizerTextKind};
+use super::FullTextIndex;
+use super::full_text_index_read::{FullTextIndexRead, PayloadMatchQueryType};
+use super::inverted_index::{ParsedQuery, TokenId};
+use super::tokenizers::Tokenizer;
 use crate::common::operation_error::OperationResult;
-use crate::index::field_index::{CardinalityEstimation, PayloadBlockCondition, ValueIndexer};
+use crate::index::field_index::{
+    CardinalityEstimation, PayloadBlockCondition, PayloadFieldIndexRead,
+};
 use crate::index::payload_config::StorageType;
 use crate::index::query_optimization::optimized_filter::ConditionCheckerFn;
-use crate::telemetry::PayloadIndexTelemetry;
 use crate::types::{
     FieldCondition, Match, MatchAny, MatchExcept, MatchPhrase, MatchText, MatchTextAny, MatchValue,
     PayloadKeyType,
 };
 
-/// Shared read surface for the writable [`FullTextIndex`] enum and the
-/// read-only `ReadOnlyFullTextIndex<S>` skeleton. Lets the
-/// [`PayloadFieldIndexRead`][crate::index::field_index::PayloadFieldIndexRead]
-/// bodies live in this module as free functions instead of being duplicated.
-///
-/// Object safety is **not** required — `for_each_token_id` is generic over
-/// `U: UserData` and `f: impl FnMut(..)`, so callers parameterize with
-/// `T: FullTextIndexRead` rather than `&dyn FullTextIndexRead`.
-pub trait FullTextIndexRead {
-    fn tokenizer(&self) -> &Tokenizer;
-    fn telemetry_index_type(&self) -> &'static str;
-
-    /// Telemetry shared between [`FullTextIndex`] and `ReadOnlyFullTextIndex<S>`.
-    /// Full-text indexes track a single per-point count, so `points_values_count`
-    /// and `points_count` are both reported as [`Self::points_count`].
-    ///
-    /// [`FullTextIndex`]: super::text_index::FullTextIndex
-    fn get_telemetry_data(&self) -> PayloadIndexTelemetry {
-        PayloadIndexTelemetry {
-            field_name: None,
-            index_type: self.telemetry_index_type(),
-            points_values_count: self.points_count(),
-            points_count: self.points_count(),
-            histogram_bucket_size: None,
+impl FullTextIndexRead for FullTextIndex {
+    fn tokenizer(&self) -> &Tokenizer {
+        match self {
+            Self::Mutable(index) => index.tokenizer(),
+            Self::Immutable(index) => index.tokenizer(),
+            Self::Mmap(index) => index.tokenizer(),
         }
     }
 
-    fn points_count(&self) -> usize;
-    fn values_count(&self, point_id: PointOffsetType) -> usize;
-    fn values_is_empty(&self, point_id: PointOffsetType) -> bool;
+    fn telemetry_index_type(&self) -> &'static str {
+        match self {
+            Self::Mutable(index) => index.telemetry_index_type(),
+            Self::Immutable(index) => index.telemetry_index_type(),
+            Self::Mmap(index) => index.telemetry_index_type(),
+        }
+    }
+
+    fn points_count(&self) -> usize {
+        match self {
+            Self::Mutable(index) => index.points_count(),
+            Self::Immutable(index) => index.points_count(),
+            Self::Mmap(index) => index.points_count(),
+        }
+    }
+
+    fn values_count(&self, point_id: PointOffsetType) -> usize {
+        match self {
+            Self::Mutable(index) => index.values_count(point_id),
+            Self::Immutable(index) => index.values_count(point_id),
+            Self::Mmap(index) => index.values_count(point_id),
+        }
+    }
+
+    fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
+        match self {
+            Self::Mutable(index) => index.values_is_empty(point_id),
+            Self::Immutable(index) => index.values_is_empty(point_id),
+            Self::Mmap(index) => index.values_is_empty(point_id),
+        }
+    }
 
     fn for_each_token_id<'a, U: UserData>(
         &self,
         iter: impl Iterator<Item = (U, &'a str)>,
         hw_counter: &HardwareCounterCell,
         f: impl FnMut(U, Option<TokenId>),
-    ) -> OperationResult<()>;
+    ) -> OperationResult<()> {
+        match self {
+            Self::Mutable(index) => index.for_each_token_id(iter, hw_counter, f),
+            Self::Immutable(index) => index.for_each_token_id(iter, hw_counter, f),
+            Self::Mmap(index) => index.for_each_token_id(iter, hw_counter, f),
+        }
+    }
 
     fn filter_query<'a>(
         &'a self,
         query: ParsedQuery,
         hw_counter: &'a HardwareCounterCell,
-    ) -> OperationResult<Box<dyn Iterator<Item = PointOffsetType> + 'a>>;
+    ) -> OperationResult<Box<dyn Iterator<Item = PointOffsetType> + 'a>> {
+        match self {
+            Self::Mutable(index) => index.filter_query(query, hw_counter),
+            Self::Immutable(index) => index.filter_query(query, hw_counter),
+            Self::Mmap(index) => index.filter_query(query, hw_counter),
+        }
+    }
 
     fn estimate_query_cardinality(
         &self,
         query: &ParsedQuery,
         condition: &FieldCondition,
         hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<CardinalityEstimation>;
+    ) -> OperationResult<CardinalityEstimation> {
+        match self {
+            Self::Mutable(index) => index.estimate_query_cardinality(query, condition, hw_counter),
+            Self::Immutable(index) => {
+                index.estimate_query_cardinality(query, condition, hw_counter)
+            }
+            Self::Mmap(index) => index.estimate_query_cardinality(query, condition, hw_counter),
+        }
+    }
 
-    fn check_match(&self, query: &ParsedQuery, point_id: PointOffsetType) -> OperationResult<bool>;
+    fn check_match(&self, query: &ParsedQuery, point_id: PointOffsetType) -> OperationResult<bool> {
+        match self {
+            Self::Mutable(index) => index.check_match(query, point_id),
+            Self::Immutable(index) => index.check_match(query, point_id),
+            Self::Mmap(index) => index.check_match(query, point_id),
+        }
+    }
 
-    /// Walk the inverted-index vocab and emit one [`PayloadBlockCondition`] per
-    /// token with at least `threshold` postings. Used to seed payload-block
-    /// scans for full-text indexes.
     fn for_each_payload_block_inner(
         &self,
         threshold: usize,
         key: PayloadKeyType,
         f: &mut dyn FnMut(PayloadBlockCondition) -> OperationResult<()>,
-    ) -> OperationResult<()>;
-
-    fn get_storage_type(&self) -> StorageType;
-
-    fn ram_usage_bytes(&self) -> usize;
-
-    fn is_on_disk(&self) -> bool;
-
-    /// Parse as [`TokenizerTextKind::Document`] and return [`ParsedQuery::Phrase`].
-    /// Returns [`None`] if there are any unseen tokens.
-    fn parse_phrase_query(
-        &self,
-        phrase: &str,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<Option<ParsedQuery>> {
-        let document = self.parse_document(phrase, hw_counter)?;
-        Ok(document.map(ParsedQuery::Phrase))
-    }
-
-    /// Parse as [`TokenizerTextKind::Query`] and return [`ParsedQuery::AllTokens`].
-    /// Returns [`None`] if there are any unseen tokens.
-    fn parse_text_query(
-        &self,
-        text: &str,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<Option<ParsedQuery>> {
-        let tokenset: Option<TokenSet> = self
-            .resolve_tokens(TokenizerTextKind::Query, text, hw_counter)?
-            .into_values()
-            .collect::<Option<TokenSet>>();
-        Ok(tokenset.map(ParsedQuery::AllTokens))
-    }
-
-    /// Parse as [`TokenizerTextKind::Query`] and return [`ParsedQuery::AnyTokens`].
-    /// Unseen tokens are ignored. Never returns [`None`].
-    fn parse_text_any_query(
-        &self,
-        text: &str,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<Option<ParsedQuery>> {
-        let tokenset = self.parse_tokenset(TokenizerTextKind::Query, text, hw_counter)?;
-        Ok(Some(ParsedQuery::AnyTokens(tokenset)))
-    }
-
-    /// Parse as provided [`TokenizerTextKind`] and return [`TokenSet`].
-    /// Unseen tokens are ignored.
-    fn parse_tokenset(
-        &self,
-        kind: TokenizerTextKind,
-        text: &str,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<TokenSet> {
-        let token_ids = self.resolve_tokens(kind, text, hw_counter)?.into_values();
-        Ok(token_ids.flatten().collect())
-    }
-
-    /// Tokenize the `text` and return a map of token -> token_id.
-    /// Missing tokens will have [`None`] as token_id.
-    fn resolve_tokens<'a>(
-        &self,
-        kind: TokenizerTextKind,
-        text: &'a str,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<AHashMap<Cow<'a, str>, Option<TokenId>>> {
-        let mut token_map = AHashMap::new();
-        self.tokenizer().tokenize(kind, text, |token| {
-            token_map.insert(token, None);
-        });
-        let iter = token_map
-            .iter_mut()
-            .map(|(token, cell)| (cell, token.as_ref()));
-        self.for_each_token_id(iter, hw_counter, |cell, token_id| *cell = token_id)?;
-        Ok(token_map)
-    }
-
-    /// Parse as [`TokenizerTextKind::Document`] and return a [`Document`].
-    /// Returns [`None`] if there are any unseen tokens.
-    fn parse_document(
-        &self,
-        text: &str,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<Option<Document>> {
-        let mut document_tokens = Vec::new();
-        let token_map = self.resolve_tokens(TokenizerTextKind::Document, text, hw_counter)?;
-        if token_map.values().any(|token_id| token_id.is_none()) {
-            return Ok(None);
+    ) -> OperationResult<()> {
+        match self {
+            Self::Mutable(index) => index.for_each_payload_block_inner(threshold, key, f),
+            Self::Immutable(index) => index.for_each_payload_block_inner(threshold, key, f),
+            Self::Mmap(index) => index.for_each_payload_block_inner(threshold, key, f),
         }
-
-        self.tokenizer()
-            .tokenize(TokenizerTextKind::Document, text, |token| {
-                let token_id = token_map
-                    .get(&token)
-                    .expect("token should be in map")
-                    .expect("token_id should be set for all tokens");
-                document_tokens.push(token_id);
-            });
-
-        Ok(Some(Document::new(document_tokens)))
     }
 
-    /// Checks the text directly against the payload value using the
-    /// full-text index tokenizer.
-    ///
-    /// `query_type` selects the parsing / matching strategy:
-    /// - `Text`    — all query tokens must appear in the document
-    /// - `Phrase`  — all query tokens must appear in exact order
-    /// - `TextAny` — at least one query token must appear
-    fn check_payload_match(
+    fn get_storage_type(&self) -> StorageType {
+        match self {
+            Self::Mutable(index) => FullTextIndexRead::get_storage_type(index),
+            Self::Immutable(index) => FullTextIndexRead::get_storage_type(index),
+            Self::Mmap(index) => FullTextIndexRead::get_storage_type(index.as_ref()),
+        }
+    }
+
+    fn ram_usage_bytes(&self) -> usize {
+        match self {
+            Self::Mutable(index) => FullTextIndexRead::ram_usage_bytes(index),
+            Self::Immutable(index) => FullTextIndexRead::ram_usage_bytes(index),
+            Self::Mmap(index) => FullTextIndexRead::ram_usage_bytes(index.as_ref()),
+        }
+    }
+
+    fn is_on_disk(&self) -> bool {
+        match self {
+            Self::Mutable(index) => FullTextIndexRead::is_on_disk(index),
+            Self::Immutable(index) => FullTextIndexRead::is_on_disk(index),
+            Self::Mmap(index) => FullTextIndexRead::is_on_disk(index.as_ref()),
+        }
+    }
+}
+
+impl PayloadFieldIndexRead for FullTextIndex {
+    fn count_indexed_points(&self) -> usize {
+        FullTextIndexRead::points_count(self)
+    }
+
+    fn filter<'a>(
+        &'a self,
+        condition: &'a FieldCondition,
+        hw_counter: &'a HardwareCounterCell,
+    ) -> OperationResult<Option<Box<dyn Iterator<Item = PointOffsetType> + 'a>>> {
+        filter(self, condition, hw_counter)
+    }
+
+    fn estimate_cardinality(
         &self,
-        payload_value: &serde_json::Value,
-        text: &str,
-        query_type: PayloadMatchQueryType,
+        condition: &FieldCondition,
         hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<bool> {
-        let query_opt = match query_type {
-            PayloadMatchQueryType::Text => self.parse_text_query(text, hw_counter)?,
-            PayloadMatchQueryType::Phrase => self.parse_phrase_query(text, hw_counter)?,
-            PayloadMatchQueryType::TextAny => self.parse_text_any_query(text, hw_counter)?,
-        };
+    ) -> OperationResult<Option<CardinalityEstimation>> {
+        estimate_cardinality(self, condition, hw_counter)
+    }
 
-        let Some(query) = query_opt else {
-            return Ok(false);
-        };
+    fn for_each_payload_block(
+        &self,
+        threshold: usize,
+        key: PayloadKeyType,
+        f: &mut dyn FnMut(PayloadBlockCondition) -> OperationResult<()>,
+    ) -> OperationResult<()> {
+        for_each_payload_block(self, threshold, key, f)
+    }
 
-        <super::text_index::FullTextIndex as ValueIndexer>::get_values(payload_value)
-            .iter()
-            .try_any(|value| match &query {
-                ParsedQuery::AllTokens(query) => {
-                    let tokenset =
-                        self.parse_tokenset(TokenizerTextKind::Document, value, hw_counter)?;
-                    Ok(tokenset.has_subset(query))
-                }
-                ParsedQuery::Phrase(query) => {
-                    let document = self.parse_document(value, hw_counter)?;
-                    Ok(document.is_some_and(|doc| doc.has_phrase(query)))
-                }
-                ParsedQuery::AnyTokens(query) => {
-                    let tokenset =
-                        self.parse_tokenset(TokenizerTextKind::Document, value, hw_counter)?;
-                    Ok(tokenset.has_any(query))
-                }
-            })
+    fn condition_checker<'a>(
+        &'a self,
+        condition: &FieldCondition,
+        hw_acc: HwMeasurementAcc,
+    ) -> Option<ConditionCheckerFn<'a>> {
+        condition_checker(self, condition, hw_acc)
+    }
+
+    fn special_check_condition(
+        &self,
+        condition: &FieldCondition,
+        payload_value: &Value,
+        hw_counter: &HardwareCounterCell,
+    ) -> OperationResult<Option<bool>> {
+        special_check_condition(self, condition, payload_value, hw_counter)
+    }
+}
+
+impl FullTextIndex {
+    #[cfg(test)]
+    pub fn query<'a>(
+        &'a self,
+        query: &'a str,
+        hw_counter: &'a HardwareCounterCell,
+    ) -> OperationResult<Box<dyn Iterator<Item = PointOffsetType> + 'a>> {
+        let Some(parsed_query) = self.parse_text_query(query, hw_counter)? else {
+            return Ok(Box::new(std::iter::empty()));
+        };
+        self.filter_query(parsed_query, hw_counter)
     }
 }
 
 /// Body for [`PayloadFieldIndexRead::filter`]. Shared between [`FullTextIndex`]
 /// and `ReadOnlyFullTextIndex<S>`.
-///
-/// [`FullTextIndex`]: super::text_index::FullTextIndex
 pub fn filter<'a, T: FullTextIndexRead>(
     index: &'a T,
     condition: &FieldCondition,
