+++ lib/segment/src/index/field_index/index_selector.rs
@@ -18,7 +18,7 @@ use crate::common::operation_error::{OperationError, OperationResult};
 use crate::data_types::index::TextIndexParams;
 use crate::id_tracker::{IdTrackerEnum, IdTrackerRead};
 use crate::index::field_index::FieldIndex;
-use crate::index::field_index::full_text_index::text_index::FullTextIndex;
+use crate::index::field_index::full_text_index::FullTextIndex;
 use crate::index::field_index::geo_index::GeoMapIndex;
 use crate::index::field_index::null_index::MutableNullIndex;
 use crate::index::field_index::numeric_index::NumericIndex;
